# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "ConversationOut": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "machine_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Machine Id"
          },
          "machine_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Machine Label"
          },
          "started_at": {
            "title": "Started At",
            "type": "string"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "machine_id",
          "machine_label",
          "title",
          "started_at",
          "updated_at"
        ],
        "title": "ConversationOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SetMachineRequest": {
        "properties": {
          "machine_id": {
            "title": "Machine Id",
            "type": "integer"
          }
        },
        "required": [
          "machine_id"
        ],
        "title": "SetMachineRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "app__api__routes_admin__MachineOut": {
        "properties": {
          "family": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Family"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "manufacturer": {
            "title": "Manufacturer",
            "type": "string"
          },
          "model_name": {
            "title": "Model Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "manufacturer",
          "model_name",
          "family"
        ],
        "title": "MachineOut",
        "type": "object"
      },
...
      "app__api__routes_machines__MachineOut": {
        "properties": {
          "document_count": {
            "title": "Document Count",
            "type": "integer"
          },
          "family": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Family"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_favorite": {
            "default": false,
            "title": "Is Favorite",
            "type": "boolean"
          },
          "last_used_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          },
          "machine_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Machine Type"
          },
          "manufacturer": {
            "title": "Manufacturer",
            "type": "string"
          },
          "model_name": {
            "title": "Model Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "manufacturer",
          "model_name",
          "family",
          "machine_type",
          "document_count"
        ],
        "title": "MachineOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Technician Manual Assistant",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/admin/machines": {
      "get": {
        "description": "Every machine in the catalog, unlike GET /api/machines which only\nreturns machines that already have a document linked \u2014 an admin fixing a\ndocument's association needs to be able to pick a machine that has zero\n(or wrong) links today.",
        "operationId": "list_all_machines_api_admin_machines_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "tma_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tma Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/app__api__routes_admin__MachineOut"
                  },
                  "title": "Response List All Machines Api Admin Machines Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List All Machines",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/api/conversations/{conversation_id}/machine": {
      "post": {
        "description": "The ONLY way a conversation's machine is set once clarification is\nneeded, or changed later (\"Change machine\"). This is always an explicit,\nconfirmed technician action -- never inferred from a later message body,\nwhich is what let a conversation's machine silently drift in the reviewed\nversion (concern #5/#6).\n\nIf a clarifying question is pending (the technician asked something\nbefore the machine was known), confirming the machine here resumes and\nanswers that ORIGINAL stored question -- it does not require the caller\nto resubmit it as a new user turn (P1-8). The resumed answer is\ngenerated and persisted as usual; this endpoint's own response stays\nConversationOut either way, so the caller reloads\nGET /conversations/{id}/messages to see it, the same as after any other\nanswer.",
        "operationId": "set_conversation_machine_api_conversations__conversation_id__machine_post",
        "parameters": [
          {
            "in": "path",
            "name": "conversation_id",
            "required": true,
            "schema": {
              "title": "Conversation Id",
              "type": "integer"
            }
          },
          {
            "in": "cookie",
            "name": "tma_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tma Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetMachineRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Conversation Machine",
        "tags": [
          "chat"
        ]
      }
    },
...
    "/api/machines": {
      "get": {
        "description": "Autocomplete search across model name, family, and manufacturer. Only\nmachines that actually have at least one indexed, approved, current\ndocument (with an approved link) are returned -- otherwise the picker\nwould offer a machine that then dead-ends into \"no manuals\" the moment\nretrieval applies its own approval/revision filters (independent\nfollow-up review P0-6, P1-6). The eligibility rules here are deliberately\nthe same set retrieval enforces, including the current-revision rule\n(P1-11): a machine whose only manual is superseded has nothing\nretrievable and must not appear.\n\nRequires auth: the equipment catalog is proprietary to the deployment\n(concern #20 -- this endpoint leaked it to unauthenticated requests).",
        "operationId": "search_machines_api_machines_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "cookie",
            "name": "tma_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tma Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/app__api__routes_machines__MachineOut"
                  },
                  "title": "Response Search Machines Api Machines Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Machines",
        "tags": [
          "machines"
        ]
      }
    },
    "/api/machines/recent": {
      "get": {
        "description": "Applies the exact same eligibility rules and `HAVING document_count > 0`\nas `search_machines()` above -- a machine a technician favorited or\nrecently used, whose only manual has since been deactivated/unapproved/\nsuperseded, is dropped rather than shown with `document_count: 0`\n(independent follow-up review P1-6's \"apply the rules consistently to\nsearch and recent machines\": found, during this pass, that this endpoint\nwas missing the `HAVING` clause `search_machines()` already had, so a\ndead manual could resurface here even though the picker correctly hid\nit). The tradeoff is explicit: a favorited-but-now-empty machine\ndisappears from recents instead of dead-ending into \"no manuals\" --\nconsistent with what the picker already does, not a new UX decision.",
        "operationId": "recent_machines_api_machines_recent_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "cookie",
            "name": "tma_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tma Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/app__api__routes_machines__MachineOut"
                  },
                  "title": "Response Recent Machines Api Machines Recent Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Recent Machines",
        "tags": [
          "machines"
        ]
      }
    },
    "/api/machines/{machine_id}/favorite": {
      "post": {
        "operationId": "set_favorite_api_machines__machine_id__favorite_post",
        "parameters": [
          {
            "in": "path",
            "name": "machine_id",
            "required": true,
            "schema": {
              "title": "Machine Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "favorite",
            "required": false,
            "schema": {
              "default": true,
              "title": "Favorite",
              "type": "boolean"
            }
          },
          {
            "in": "cookie",
            "name": "tma_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tma Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Favorite",
        "tags": [
          "machines"
        ]
      }
    },
    "/api/machines/{machine_id}/touch": {
      "post": {
        "operationId": "touch_recent_api_machines__machine_id__touch_post",
        "parameters": [
          {
            "in": "path",
            "name": "machine_id",
            "required": true,
            "schema": {
              "title": "Machine Id",
              "type": "integer"
            }
          },
          {
            "in": "cookie",
            "name": "tma_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tma Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Touch Recent",
        "tags": [
          "machines"
        ]
      }
    },
...
  }
}